adds userrank
