+++ src/components/user-rating.js
@@ -1,32 +1,30 @@
-import AbstractComponent from "./abstract-component";
+import AbstractSmartComponent from "./abstract-smart-component";
+import {getUserRank} from "../util";
 
-const getUserRank = (rank) => {
-  if (rank >= 1 && rank <= 10) {
-    return `novice`;
-  } else if (rank >= 11 && rank <= 20) {
-    return `fan`;
-  } else if (rank >= 21) {
-    return `movie buff`;
-  }
-  return ``;
-};
-
-const createUserRatingTemplate = (data) => {
-  const rank = getUserRank(data);
+const createUserRatingTemplate = (moviesModel) => {
+  const rank = moviesModel.getCardsAll().filter((it) => it.watched).length;
   return (`<section class="header__profile profile">
-    <p class="profile__rating">${rank}</p>
+    <p class="profile__rating">${getUserRank(rank)}</p>
     <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
   </section>`);
 };
 
-export default class UserRating extends AbstractComponent {
-  constructor(data) {
+export default class UserRating extends AbstractSmartComponent {
+  constructor(moviesModel) {
     super();
 
-    this._data = data;
+    this._moviesModel = moviesModel;
+    this._onDataChange = this._onDataChange.bind(this);
+    this._moviesModel.setDataChangeHandler(this._onDataChange);
   }
 
   getTemplate() {
-    return createUserRatingTemplate(this._data);
+    return createUserRatingTemplate(this._moviesModel);
   }
+
+  _onDataChange() {
+    this.rerender();
+  }
+
+  recoveryListeners() {}
 }
